fix: исправить модификаторы доступа

+++ main-service/src/main/java/ru/yandex/practicum/compilation/dto/UpdateCompilationRequest.java
@@ -11,7 +11,7 @@ import java.util.Set;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UpdateCompilationRequest {
-    private Set<Integer> events;
+    private Set<Long> events;
     private Boolean pinned;
     @Size(max = 50)
     private String title;
